docs 결제시 각종 테이블 업데이트
subject : 결제시 각종 테이블 업데이트

+++ So1omon/src/main/java/com/kh/so1omon/product/model/vo/Order.java
@@ -17,7 +17,7 @@ public class Order {
 	private long orderNo;
 	private String userId;
 	private long userNo;
-	private int tracking;
+	private long tracking;
 	private Date orderDate;
 	private String cashType;
 	private String status;
